🧹 Tidy: Standardize attribute setters and deduplicate validation rules
- Refactor `ChurchServiceItem` attribute setters to use short arrow functions and the `filled()` helper.
- Extract common "trimmed text" validation logic into a reusable `App\Rules\TrimmedText` rule.
- Update `Meeting`, `SongAuthor`, and `SongBook` models to use the centralized `TrimmedText` rule.
- Align PHPDoc array shape annotations with model nullability for the `description` field in presenters and DTOs to resolve PHPStan Level 8 errors.
