+++ app/Data/PublicMeetingReadModel.php
@@ -30,7 +30,7 @@ final readonly class PublicMeetingReadModel
     ) {}
 
     /**
-     * @param  Collection<int, array{area: string, description: string, heading: string, image_url: string, slug: string, url: string}>|array<int, array{area: string, description: string, heading: string, image_url: string, slug: string, url: string}>  $links
+     * @param  Collection<int, array{area: string, description: string|null, heading: string, image_url: string, slug: string, url: string}>|array<int, array{area: string, description: string|null, heading: string, image_url: string, slug: string, url: string}>  $links
      * @param  Collection<int, mixed>  $pastEvents
      * @return array{
      *     area: string,
@@ -40,7 +40,7 @@ final readonly class PublicMeetingReadModel
      *     headingpicture: ?string,
      *     headingpictureMobile: ?string,
      *     headingpictureTablet: ?string,
-     *     links: Collection<int, array{area: string, description: string, heading: string, image_url: string, slug: string, url: string}>,
+     *     links: Collection<int, array{area: string, description: string|null, heading: string, image_url: string, slug: string, url: string}>,
      *     meeting: Meeting,
      *     metaDescription: string,
      *     page: ?Page,
