+++ app/Data/PublicPageReadModel.php
@@ -22,7 +22,7 @@ final readonly class PublicPageReadModel
     ) {}
 
     /**
-     * @param  Collection<int, array{area: string, description: string, heading: string, image_url: string, slug: string, url: string}>|array<int, array{area: string, description: string, heading: string, image_url: string, slug: string, url: string}>  $links
+     * @param  Collection<int, array{area: string, description: string|null, heading: string, image_url: string, slug: string, url: string}>|array<int, array{area: string, description: string|null, heading: string, image_url: string, slug: string, url: string}>  $links
      * @return array{
      *     area: string,
      *     content: string,
@@ -31,7 +31,7 @@ final readonly class PublicPageReadModel
      *     headingpicture: ?string,
      *     headingpictureMobile: ?string,
      *     headingpictureTablet: ?string,
-     *     links: Collection<int, array{area: string, description: string, heading: string, image_url: string, slug: string, url: string}>,
+     *     links: Collection<int, array{area: string, description: string|null, heading: string, image_url: string, slug: string, url: string}>,
      *     metaDescription: string,
      *     page: ?Page,
      *     slug: string
